Leonardo Board.
Flashing under Linux with avra requires changes in the avrdude.conf file. Detail
are at http://savannah.nongnu.org/bugs/?36901. Otherwise the EEPROm does not get anything useful
Windows with the AVR Studio works. 

The USART used for the terminal is the real USART as labled on the board, not the USB CDC.

git-svn-id: svn://svn.code.sf.net/p/amforth/code@1353 fd1c0b69-5d68-47f5-910c-86182ebb6ecd